feat(cloudflare): move deploy command to cloudflare package
