+++ examples/tpr-demo/app/components/tpr-visualization.tsx
@@ -574,7 +574,7 @@ export default function TPRVisualization() {
 
           <div className="code-block">
             <span className="dimmed">$ </span>
-            <span className="highlight">vinext deploy --experimental-tpr</span>
+            <span className="highlight">npx @vinext/cloudflare deploy --experimental-tpr</span>
             {"\n\n"}
             <span className="dimmed">{"  "}Project: tpr-demo</span>
             {"\n"}
