+++ packages/cloudflare/package.json
@@ -8,11 +8,30 @@
     "url": "git+https://github.com/cloudflare/vinext.git",
     "directory": "packages/cloudflare"
   },
+  "bin": {
+    "vinext-cloudflare": "dist/cli.js"
+  },
   "files": [
     "dist"
   ],
   "type": "module",
   "exports": {
+    "./internal/deploy": {
+      "types": "./dist/deploy.d.ts",
+      "import": "./dist/deploy.js"
+    },
+    "./internal/deploy-help": {
+      "types": "./dist/deploy-help.d.ts",
+      "import": "./dist/deploy-help.js"
+    },
+    "./internal/deploy-config": {
+      "types": "./dist/deploy-config.d.ts",
+      "import": "./dist/deploy-config.js"
+    },
+    "./internal/tpr": {
+      "types": "./dist/tpr.d.ts",
+      "import": "./dist/tpr.js"
+    },
     "./cache/*": {
       "types": "./dist/cache/*.d.ts",
       "import": "./dist/cache/*.js"
