+++ packages/cloudflare/src/deploy-help.ts
@@ -1,0 +1,51 @@
+export function formatDeployHelp(): string {
+  return `
+  @vinext/cloudflare deploy - Deploy to Cloudflare Workers
+
+  Usage: @vinext/cloudflare deploy [options]
+
+  One-command deployment to Cloudflare Workers. Automatically:
+    - Detects App Router or Pages Router
+    - Validates setup from vinext init --platform=cloudflare
+    - Builds the project with Vite
+    - Deploys via wrangler
+
+  Options:
+    --preview                Deploy to preview environment (same as --env preview)
+    --env <name>             Deploy using wrangler env.<name>
+    --name <name>            Custom Worker name (default: from package.json)
+    --skip-build             Skip the build step (use existing dist/)
+    --dry-run                Validate setup without building or deploying
+    --prerender-all          Pre-render discovered routes after building (future
+                             releases will auto-populate the remote cache)
+    --prerender-concurrency <count>
+                             Maximum number of routes to pre-render in parallel
+    -h, --help               Show this help
+
+  Experimental:
+    --experimental-tpr               Enable Traffic-aware Pre-Rendering
+    --tpr-coverage <pct>             Traffic coverage target, 0-100 (default: 90)
+    --tpr-limit <count>              Hard cap on pages to pre-render (default: 1000)
+    --tpr-window <hours>             Analytics lookback window in hours (default: 24)
+
+  TPR (Traffic-aware Pre-Rendering) uses Cloudflare zone analytics to determine
+  which pages get the most traffic and pre-renders them into KV cache during
+  deploy. This feature is experimental and must be explicitly enabled. Requires
+  a custom domain (zone analytics are unavailable on *.workers.dev) and the
+  CLOUDFLARE_API_TOKEN environment variable with Zone.Analytics read permission.
+
+  Examples:
+    npx @vinext/cloudflare deploy                              Build and deploy to production
+    npx @vinext/cloudflare deploy --preview                    Deploy to a preview URL
+    npx @vinext/cloudflare deploy --env staging                Deploy using wrangler env.staging
+    npx @vinext/cloudflare deploy --dry-run                    Validate setup without building or deploying
+    npx @vinext/cloudflare deploy --name my-app                Deploy with a custom Worker name
+    npx @vinext/cloudflare deploy --experimental-tpr           Enable TPR during deploy
+    npx @vinext/cloudflare deploy --experimental-tpr --tpr-coverage 95   Cover 95% of traffic
+    npx @vinext/cloudflare deploy --experimental-tpr --tpr-limit 500     Cap at 500 pages
+`;
+}
+
+export function printDeployHelp(): void {
+  console.log(formatDeployHelp());
+}
